+++ src/App.scss
@@ -10,3 +10,26 @@
     display: flex;
   }
 }
+
+.center-content {
+  text-align: center;
+  align-content: center;
+}
+
+.border {
+  &-red {
+    border: 1px solid red;
+  }
+
+  &-blue {
+    border: 1px solid blue;
+  }
+
+  &-green {
+    border: 1px solid green;
+  }
+
+  &-black {
+    border: 1px solid black;
+  }
+}
