+++ src/game/systems/Systems.ts
@@ -1,5 +1,5 @@
 import type { Ticker } from 'pixi.js';
-import type { Entity } from '../../../utils/ecsUtils';
+import type { Entity } from '../utils/ecsUtils';
 
 export type System = {
   update: (entities: Entity[], time: Ticker) => void;
