+++ src/game/utils/mapUtils.ts
