+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App';
-import './index.css';
+import '../index.css';
 import {
   BrowserRouter,
 } from 'react-router-dom';
