+++ src/react/components/Game.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { SidePanel } from './SidePanel';
-import { GameMap } from './GameMap/GameMap';
+import { GameMap } from './GameMap';
 
 export const Game: React.FC = () => {
   return (
