+++ src/react/components/GameMap.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import './GameMap.scss';
-import { PixiStage } from '../../canvas/PixiStage';
+import { PixiStage } from '../../game/PixiStage';
 
 export const GameMap: React.FC = () => {
   return <PixiStage />;
