+++ src/react/components/SidePanel.tsx
