--- src/utils/common.scss
@@ -1,22 +1,0 @@
-.center-content {
-  text-align: center;
-  align-content: center;
-}
-
-.border {
-  &-red {
-    border: 1px solid red;
-  }
-
-  &-blue {
-    border: 1px solid blue;
-  }
-
-  &-green {
-    border: 1px solid green;
-  }
-
-  &-black {
-    border: 1px solid black;
-  }
-}
